requirements.txt
flask==2.3.2
flask_cors==3.0.10
pypdf==3.9.1
langchain==0.0.198
flask_sqlalchemy==3.0.3
sqlalchemy_utils==0.41.1
openai==0.27.8
tiktoken==0.4.0
faiss-cpu==1.7.4
rq==1.15.0
flask_httpauth==4.8.0
python-dotenv==1.0.0
prometheus-client==0.17.0
slack_sdk==3.21.3
flask_restx==1.1.0
beautifulsoup4==4.12.2